+++ f_o.c
@@ -6,7 +6,7 @@
 /*   By: etuffleb <etuffleb@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/20 22:30:06 by etuffleb          #+#    #+#             */
-/*   Updated: 2019/01/23 19:15:50 by etuffleb         ###   ########.fr       */
+/*   Updated: 2019/01/23 19:56:36 by etuffleb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 char	*f_o(va_list ap)
 {
-	unsigned int	integer;
+	long long int	integer;
 	char			*int_num;
 
-	integer = va_arg(ap, unsigned int);
+	integer = va_arg(ap, long long int);
 	int_num = ft_itoa_base(integer, 8);
 	return (int_num);
 }
